move tooltip.js back to the component
